Vector spaces

Vector spaces

The type hierarchy for representing vector spaces

TensorKit.VectorSpace — Type.

abstract type VectorSpace end

Abstract type at the top of the type hierarchy for denoting vector spaces.

source

TensorKit.ElementarySpace — Type.

abstract type ElementarySpace{𝕜} <: VectorSpace end

Elementary finite-dimensional vector space over a field 𝕜 that can be used as the index space corresponding to the indices of a tensor.

Every elementary vector space should respond to the methods conj and dual, returning the complex conjugate space and the dual space respectively. The complex conjugate of the dual space is obtained as dual(conj(V)) === conj(dual(V)). These different spaces should be of the same type, so that a tensor can be defined as an element of a homogeneous tensor product of these spaces.

source

TensorKit.GeneralSpace — Type.

struct GeneralSpace{𝕜} <: ElementarySpace{𝕜}

A finite-dimensional space over an arbitrary field 𝕜 without additional structure. It is thus characterized by its dimension, and whether or not it is the dual and/or conjugate space. For a real field 𝕜, the space and its conjugate are the same.

source

TensorKit.InnerProductSpace — Type.

abstract type InnerProductSpace{𝕜} <: ElementarySpace{𝕜} end

Abstract type for denoting vector with an inner product and a corresponding metric, which can be used to raise or lower indices of tensors.

source

TensorKit.EuclideanSpace — Type.

abstract type EuclideanSpace{𝕜} <: InnerProductSpace{𝕜} end

Abstract type for denoting real or complex spaces with a standard (Euclidean) inner product (i.e. orthonormal basis), such that the dual space is naturally isomorphic to the conjugate space (in the complex case) or even to the space itself (in the real case), also known as the category of finite-dimensional Hilbert spaces $FdHilb$.

source

TensorKit.CartesianSpace — Type.

struct CartesianSpace <: EuclideanSpace{ℝ}

A real euclidean space ℝ^d, which is therefore self-dual. CartesianSpace has no additonal structure and is completely characterised by its dimension d. This is the vector space that is implicitly assumed in most of matrix algebra.

source

TensorKit.ComplexSpace — Type.

struct ComplexSpace <: EuclideanSpace{ℂ}

A standard complex vector space ℂ^d with Euclidean inner product and no additional structure. It is completely characterised by its dimension and whether its the normal space or its dual (which is canonically isomorphic to the conjugate space).

source

TensorKit.RepresentationSpace — Type.

abstract type RepresentationSpace{G<:Sector} <: EuclideanSpace{ℂ} end

Complex Euclidean space with a direct sum structure corresponding to different superselection sectors of type G<:Sector, e.g. the elements or irreps of a compact or finite group, or the labels of a unitary fusion category.

source

TensorKit.GenericRepresentationSpace — Type.

struct GenericRepresentationSpace{G<:Sector} <: RepresentationSpace{G}

Generic implementation of a representation space, i.e. a complex Euclidean space with a direct sum structure corresponding to different superselection sectors of type G<:Sector, e.g. the irreps of a compact or finite group, or the labels of a unitary fusion category.

source

TensorKit.ZNSpace — Type.

struct ZNSpace{N} <: AbstractRepresentationSpace{ZNIrrep{N}}

Optimized implementation of a graded ℤ_N space, i.e. a complex Euclidean space graded by the irreps of type ZNIrrep{N}.

source

TensorKit.CompositeSpace — Type.

abstract type CompositeSpace{S<:ElementarySpace} <: VectorSpace end

Abstract type for composite spaces that are defined in terms of a number of elementary vector spaces of a homogeneous type S<:ElementarySpace{𝕜}.

source

TensorKit.ProductSpace — Type.

struct ProductSpace{S<:ElementarySpace, N} <: CompositeSpace{S}

A ProductSpace is a tensor product space of N vector spaces of type S<:ElementarySpace. Only tensor products between ElementarySpace objects of the same type are allowed.

source

The type hierarchy for representing sectors

Sector
AbelianIrrep
ZNIrrep{N}
U1Irrep
SU2Irrep
CU1Irrep
FusionStyle
Abelian
NonAbelian
SimpleNonAbelian
DegenerateNonAbelian

Methods often apply similar to e.g. spaces and corresponding tensors or tensor maps, e.g.:

field
sectortype
sectors
checksectors
dim
dims
blocksectors
blockdim

The following methods act specifically ElementarySpace spaces

isdual
dual
conj
flip
:⊕
oneunit

or also on ProductSpace

fuse
:⊗
one